chore: moved MarkerFrequencyInputWidget to Marker module
